+++ src/order_service/sql/queries/orders.sql
@@ -7,7 +7,7 @@ RETURNING *;
 SELECT id, product_id FROM orders WHERE user_id=$1 LIMIT $2 OFFSET $3;
 
 -- name: GetOrderById :one
-SELECT id, created_at, modified_at, product_id FROM orders WHERE id=$1;
+SELECT id, created_at, modified_at, product_id FROM orders WHERE id=$1 AND user_id=$2;
 
 -- name: DeleteOrder :exec
-DELETE FROM orders WHERE id=$1;
+DELETE FROM orders WHERE id=$1 AND user_id=$2;
